feat: add alert when you dont have a product application on spg (FLEX-1157)
